v2 for tracker
added a team table to be able to track a new team, and toggle between them. added toggles for level and season. added the mascot. added a separation between the header filters and score and the buttons and tracking elements below. accounted for empty data as new teams are added without having uploaded rosters, schedule, etc.